task 3.1, 3.2, 3.3, 3.4
3.1. Uniform time distribution:
In this policy you should distribute the tickets in such a way, so as to achieve a uniform time allocation to the processes (assuming that your implementation of the random number generator achieves a uniform distribution of the returned values).

3.2. Priority scheduling:
This scheduling policy will take the process priority into consideration while deciding the number of tickets to allocate. For example, given two processes p1 and p2 having priorities 1 and 2 accordingly, process p2 will receive approximately twice the run-time received by p1. You should implement a new system call: void priority(int); which can be used by a process to change its priority. The priority of a new processes is 10.

3.3. Dynamic tickets allocation:
This policy will dynamically reallocate the tickets in response to the process behavior. A newly created process will get 20 tickets. Each time a process performs a blocking system call, it will receive additional 10 tickets (up to a maximum of 100 tickets) and each time a process ends the quanta without performing a blocking system call, the amount of the tickets owned by the process will be reduced by 1 (to the minimum of 1).
3.4. Changing the current policy:
In order to be able to select a desired ticket allocation policy you are required to implement a system call void policy(int). The system call receives a policy identifier (i.e., 0, 1 and 2) as an argument and changes the currently used tickets allocation policy. In addition, the system call should redistribute tickets (among the processes) according to the selected policy default values. Additionally, create a user space program called policy which must accept an argument, representing the required policy, and perform a call to the policy system call.

 The default ticket allocation policy should be uniform time distribution.